+++ db/migrations/000091_work_pricing_flags.up.sql
@@ -1,5 +1,5 @@
 ALTER TABLE work
-    MODIFY price DECIMAL(10, 2) NULL,
-    ADD COLUMN price_to DECIMAL(10, 2) NULL AFTER price,
+    MODIFY price BIGINT NULL,
+    ADD COLUMN price_to BIGINT NULL AFTER price,
     ADD COLUMN negotiable BOOLEAN NOT NULL DEFAULT false AFTER price_to,
     ADD COLUMN hide_phone BOOLEAN NOT NULL DEFAULT false AFTER negotiable;
